backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
]
